+++ src/routing.rs
@@ -1,150 +1,7 @@
-//! This module is decoupled / independent.
-
-use crate::util::ClosureNew;
-use serde::{Deserialize, Serialize};
+use crate::url::{self, Url};
+use crate::util::{self, ClosureNew};
 use std::convert::{identity, TryFrom, TryInto};
 use wasm_bindgen::{closure::Closure, JsCast, JsValue};
-
-/// Repeated here from `seed::util`, to make this module standalone.  Once we have a Gloo module
-/// that handles this, use it.
-mod util {
-    /// Convenience function to avoid repeating expect logic.
-    pub fn window() -> web_sys::Window {
-        web_sys::window().expect("Can't find the global Window")
-    }
-
-    /// Convenience function to access the `web_sys` DOM document.
-    pub fn document() -> web_sys::Document {
-        window().document().expect("Can't find document")
-    }
-
-    /// Convenience function to access `web_sys` history
-    pub fn history() -> web_sys::History {
-        window().history().expect("Can't find history")
-    }
-}
-
-/// Contains all information used in pushing and handling routes.
-/// Based on [React-Reason's router](https://github.com/reasonml/reason-react/blob/master/docs/router.md).
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
-pub struct Url {
-    pub path: Vec<String>,
-    pub search: Option<String>,
-    pub hash: Option<String>,
-    pub title: Option<String>,
-}
-
-impl Url {
-    /// Helper that ignores hash, search and title, and converts path to Strings.
-    ///
-    /// # Refenences
-    /// * [MDN docs](https://developer.mozilla.org/en-US/docs/Web/API/History_API)
-    pub fn new<T: ToString>(path: Vec<T>) -> Self {
-        Self {
-            path: path.into_iter().map(|p| p.to_string()).collect(),
-            hash: None,
-            search: None,
-            title: None,
-        }
-    }
-
-    /// Builder-pattern method for defining hash.
-    ///
-    /// # Refenences
-    /// * [MDN docs](https://developer.mozilla.org/en-US/docs/Web/API/HTMLHyperlinkElementUtils/hash)
-    pub fn hash(mut self, hash: &str) -> Self {
-        self.hash = Some(hash.into());
-        self
-    }
-
-    /// Builder-pattern method for defining search.
-    ///
-    /// # Refenences
-    /// * [MDN docs](https://developer.mozilla.org/en-US/docs/Web/API/HTMLHyperlinkElementUtils/search)
-    pub fn search(mut self, search: &str) -> Self {
-        self.search = Some(search.into());
-        self
-    }
-
-    pub fn title(mut self, title: &str) -> Self {
-        self.title = Some(title.into());
-        self
-    }
-}
-
-impl From<web_sys::Url> for Url {
-    fn from(url: web_sys::Url) -> Self {
-        let path = {
-            let mut path = url.pathname();
-            // Remove leading `/`.
-            path.remove(0);
-            path.split('/').map(ToOwned::to_owned).collect::<Vec<_>>()
-        };
-
-        let hash = {
-            let mut hash = url.hash();
-            if hash.is_empty() {
-                None
-            } else {
-                // Remove leading `#`.
-                hash.remove(0);
-                Some(hash)
-            }
-        };
-
-        let search = {
-            let mut search = url.search();
-            if search.is_empty() {
-                None
-            } else {
-                // Remove leading `?`.
-                search.remove(0);
-                Some(search)
-            }
-        };
-
-        Self {
-            path,
-            hash,
-            search,
-            title: None,
-        }
-    }
-}
-
-impl TryFrom<String> for Url {
-    type Error = String;
-
-    fn try_from(relative_url: String) -> Result<Self, Self::Error> {
-        let dummy_base_url = "http://example.com";
-        web_sys::Url::new_with_base(&relative_url, dummy_base_url)
-            .map(Url::from)
-            .map_err(|_| format!("`{}` is invalid relative URL", relative_url))
-    }
-}
-
-impl From<Vec<String>> for Url {
-    fn from(path: Vec<String>) -> Self {
-        Url::new(path)
-    }
-}
-
-// todo: Do we need both from Vec<&str> and Vec<String> ?
-impl From<Vec<&str>> for Url {
-    fn from(path: Vec<&str>) -> Self {
-        Url::new(path)
-    }
-}
-
-/// For setting up landing page routing. Unlike normal routing, we can't rely
-/// on the popstate state, so must go off path, hash, and search directly.
-pub fn current_url() -> Url {
-    let current_url = util::window().location().href().expect("get `href`");
-
-    web_sys::Url::new(&current_url)
-        .expect("create `web_sys::Url` from the current URL")
-        .into()
-}
 
 /// Add a new route using history's `push_state` method.
 ///
@@ -197,7 +54,7 @@ pub fn setup_popstate_listener<Ms>(
                 serde_json::from_str(&state_str).expect("Problem deserializing popstate state")
             }
             // Only update when requested for an update by the user.
-            None => current_url(),
+            None => url::current(),
         };
 
         if let Some(routing_msg) = routes(url) {
